requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.16.0
idna==3.11
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
setuptools==82.0.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
